requirements.txt
# requirements.txt

# Core Libraries
playwright
llm_utils @ git+https://github.com/OzMaatuk/LLMUtils.git

# Configuration
python-dotenv
configparser

# Testing
pytest
pytest-mock